Precompile only if files do not exist (may require explicit clean)

+++ jeuclid-core/src/main/script/precompile.groovy
@@ -33,37 +33,40 @@ File basedir = new File("${project.basedir}");
 //  log.info("Reason: "+e.getMessage());
 //}
 
-log.info("Preloading operator dictionary 2...");
 File newDict2 = new File(basedir,"target/classes/net/sourceforge/jeuclid/moDictionary.ser");
 File oldDict2 = new File(basedir,"target/classes/net/sourceforge/jeuclid/moDictionary.xml");
-ant.delete(file:newDict2);
-Object dict2 = OperatorDictionary2.getInstance();
-os = new FileOutputStream(newDict2);
-oo = new ObjectOutputStream(os);
-oo.writeObject(dict2);
-oo.close();
+if (!newDict2.exists()) {
+  log.info("Preloading operator dictionary 2...");
+  Object dict2 = OperatorDictionary2.getInstance();
+  os = new FileOutputStream(newDict2);
+  oo = new ObjectOutputStream(os);
+  oo.writeObject(dict2);
+  oo.close();
+}
 ant.delete(file:oldDict2);
 
-log.info("Preloading operator dictionary 3...");
 File newDict3 = new File(basedir,"target/classes/net/sourceforge/jeuclid/appendixc.ser");
 File oldDict3 = new File(basedir,"target/classes/net/sourceforge/jeuclid/appendixc.xml");
-ant.delete(file:newDict3);
-Object dict3 = OperatorDictionary3.getInstance();
-os = new FileOutputStream(newDict3);
-oo = new ObjectOutputStream(os);
-oo.writeObject(dict3);
-oo.close();
+if (!newDict3.exists()) {
+  log.info("Preloading operator dictionary 3...");
+  Object dict3 = OperatorDictionary3.getInstance();
+  os = new FileOutputStream(newDict3);
+  oo = new ObjectOutputStream(os);
+  oo.writeObject(dict3);
+  oo.close();
+}
 ant.delete(file:oldDict3);
 
-log.info("Preloading character mappings...");
 File newMap = new File(basedir,"target/classes/net/sourceforge/jeuclid/charmap.ser");
 File oldMap = new File(basedir,"target/classes/net/sourceforge/jeuclid/UnicodeData.txt");
-ant.delete(file:newMap);
-CharacterMapping map = CharacterMapping.getInstance();
-os = new FileOutputStream(newMap);
-oo = new ObjectOutputStream(os);
-oo.writeObject(map);
-oo.close();
+if (!newMap.exists()) {
+  log.info("Preloading character mappings...");
+  CharacterMapping map = CharacterMapping.getInstance();
+  os = new FileOutputStream(newMap);
+  oo = new ObjectOutputStream(os);
+  oo.writeObject(map);
+  oo.close();
+}
 ant.delete(file:oldMap);
 
 log.info("Precompilation done!");
